src/main/java/com/anirudhology/leetcoding/array/ContainsDuplicate.java
package com.anirudhology.leetcoding.array;

import java.util.HashSet;
import java.util.Set;

public class ContainsDuplicate {

    public boolean containsDuplicate(int[] nums) {
        // Special case
        if (nums == null || nums.length == 0) {
            throw new IllegalArgumentException("Array cannot be null or empty.");
        }
        // Set to store the elements
        Set<Integer> elements = new HashSet<>();
        // Traverse through the array
        for (int num : nums) {
            if (!elements.add(num)) {
                return true;
            }
        }
        return false;
    }
}
